+++ instrumentation/CHANGELOG.md
@@ -3,6 +3,8 @@ Change Log
 
 ## Unreleased
 
+## 1.7.0 (2025-03-01)
+
 - **This version requires (at least) android-junit5 1.12.0.0 and JUnit 5.12.0.**
 - Migrate to new TestPlan API in JUnit 5.12, which changed in a binary-incompatible fashion